% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fc_check}
\alias{fc_check}
\title{Check Italian fiscal codes}
\usage{
fc_check(fc)
}
\arguments{
\item{fc}{A character vector of fiscal codes.}
}
\value{
The function return \code{TRUE} if the fiscal code is wrong,
\code{FALSE} otherwise.
}
\description{
The function performs fiscal codes (both temporary and definitive) check,
computing the last (control) character basing on the others.  The algorithm
is intended to be quite "draconian", so you'll better make some precleaning
(remove blank spaces and upcase) if needed. See examples.
}
\examples{
## As i've said, the algo is quite draconian. Consider the code
## "QWEASD34D12H221M" (fictitious but right). I would like
## "qWeASd34D12h 221M   " to be (and it's) TRUE, because i consider it
## as an error (it's diminishing match probability).

fc_check("qWeASd34D12h 221M   ")

## So you'll betted do pre-cleaning (upcase and blank trimming)

fc <- c("qWeASd34D12h 221M   ", "12312312312")
fc <- gsub(" ","", toupper(fc))
fc_check(fc)

## Check results with (i.e.)
## http://www.nonsolocap.it/codice-fiscale/controllo/
}
\references{
Law source: D.M. (Ministry of Finance) n. 13813 - 23/12/76 -
"Sistemi di codificazione dei soggetti da iscrivere all'anagrafe
tributaria". Supp. ord. G.U. 345 29/12/1976.
}
\keyword{check}
\keyword{code}
\keyword{codice}
\keyword{codici}
\keyword{controllo}
\keyword{fiscal}
\keyword{fiscale}
\keyword{fiscali}

