% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/guess_fc.R
\name{guess_fc}
\alias{guess_fc}
\title{Guess Italian Fiscal Code}
\usage{
guess_fc(surname = NULL, name = NULL, birthdate = NULL, female = NULL,
  codice_catastale = NULL)
}
\arguments{
\item{surname}{character, surname}

\item{name}{character, names}

\item{birthdate}{Date, date of birth}

\item{female}{logical, female indicator variable (\code{FALSE} = man,
\code{TRUE} = woman)}

\item{codice_catastale}{italian 'codice catastale' (an identifier) of
the 'comune' of birth}
}
\value{
The function return a character vector of fiscal code.
}
\description{
The function tryies to guess regular fiscal code, extracting relevant
alphanumeric digits from surname, name, birth date, gender and 'codice
catastale' (computing the last character, the control digit).
}
\examples{
## ... using fake data
surnames <- c("Rossi", "Bianchi")
names <- c("Mario", "Giovanna")
birthdates <- as.Date(c("1960-01-01", "1970-01-01"))
comune_of_birth <- c("F205", # milan
                     "H501") # rome
female <- c(FALSE, TRUE)
guess_fc(surnames, names, birthdates, female, comune_of_birth)
## c("RSSMRA60A01F205T", "BNCGNN70A41H501V")
}

