% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ifctools-package.R
\docType{package}
\name{ifctools}
\alias{ifctools}
\alias{ifctools-package}
\title{Miscellaneous utils for italian fiscal code}
\description{
Miscellaneous utils for italian fiscal code
}
\details{
Fiscal codes are unsatisfactory but ubiquitous people identification
codes in Italy, especially useful for merge purposes among datasets.

These package provide tools to deal with fiscal codes.

For more infos about fiscal code, see
\url{http://en.wikipedia.org/wiki/Italian_fiscal_code_card}.

WARNING: these aim to be reasonably good routines. Unfortunately can't
be perfect, since in some rare cases of practical importance (especially
people not born in Italy) the true code could be generated in an
unpredictable way (official algorithm is not completely public)

As wikipedia puts it, "On the internet, there are several web sites and
software packages that allow one to calculate one's tax code. However,
the only official tax code is the one provided by the tax office, which
handles cases of identical tax codes (which is a pretty frequent case
for people not born in Italy, as in this case the 4-characters town code
in the codice fiscale is replace by a 3-digit country code), an
arbitrary changing of a tax code, as well as cases where a code is
incorrect, but still valid (because provided by the tax office)."
}

